// src/app/models/entidad.interface.ts
export interface Entidad {
    EntidadId:  number;
    Nombre:     string;
    Do:         boolean;
    Lu:         boolean;
    Ma:         boolean;
    Mi:         boolean;
    Ju:         boolean;
    Vi:         boolean;
    Sa:         boolean;
    Activo:     boolean;
    UserAudit:  string;
    ConfiguracionString:string;
}